feature/login-screen * split home page
