+++ src/components/btn/btn.scss
@@ -69,7 +69,7 @@
 }
 
 .btn-fb-blue {
-  background: $btn-blue;
+  background: $lightblue;
   color: $main-font-color;
   font-weight: 500;
   text-transform: none;
