+++ src/styles/_colors.scss
@@ -11,4 +11,4 @@ $btn-yellow: #fffdaf;
 $btn-lightgray: #f5f5f5;
 $font-gray: #9e9e9e;
 $btn-orange: #ffd1af;
-$btn-blue: #1877F2;
+$lightblue: #1877f2;
